% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_to_list.R
\name{data_frame_to_list}
\alias{data_frame_to_list}
\title{Utility function for data type conversion.}
\usage{
data_frame_to_list(x)
}
\arguments{
\item{x}{A data.frame with logical columns representing sets.}
}
\value{
A list of sets.
}
\description{
Utility function for data type conversion.
}
\examples{
d <- tibble(name = 1:6,
            A = c(rep(TRUE, 5), FALSE),
            B = rep(c(FALSE, TRUE), each = 3))
print(d)
data_frame_to_list(d)
}
