% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOverlap.R
\name{findOverlap}
\alias{findOverlap}
\title{find overlap between sets in a list}
\usage{
findOverlap(setlist, xlim = c(0, 1), ylim = c(0, 1))
}
\arguments{
\item{setlist}{list of character vectors.}

\item{xlim}{vector with 2 numbers, x axis limits for the venn diagram.}

\item{ylim}{vector with 2 numbers, y axis limits for the venn diagram.}
}
\value{
list with 2 items: a data.frame with information about the groups (sizes, coordinates, etc.), and a data.frame containing the x and y coordinates for the venn diagram
}
\description{
find overlap between sets in a list
}
\details{
the function uses geom_path for the dendrogram, so ... takes all arguements that geom_path would also take, such as color, size, etc.
}
\examples{
library(ggeuler)
library(ggplot2)

set1 <- rownames(subset(mtcars, mpg>18))
set2 <- rownames(subset(mtcars, qsec>18))
set3 <- rownames(subset(mtcars, cyl<5))
carset <- list(highMpg=set1, highQsec=set2, lowCyl=set3)

findOverlap(carset)

}
