% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_euler.R
\name{geom_euler}
\alias{geom_euler}
\title{create a venn diagram in ggplot2}
\usage{
geom_euler(
  setlist,
  textsize = 4,
  xlim = c(0, 1),
  ylim = c(0, 1),
  fillCol = TRUE,
  borderCol = FALSE,
  fixedCoords = TRUE,
  ...
)
}
\arguments{
\item{setlist}{list of character vectors.}

\item{textsize}{integer, defining the text size of the numbers in the graph.}

\item{xlim}{vector with 2 numbers, x axis limits for the venn diagram.}

\item{ylim}{vector with 2 numbers, y axis limits for the venn diagram.}

\item{fillCol}{boolean, should the fill color indicate the list items?}

\item{borderCol}{boolean, should the border color indicate the list items?}

\item{fixedCoords}{boolean, should coord_fixed be applied? If TRUE, makes the graph into a square.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
a ggplot2 layer objects (geom_polyon for the venn diagram) that can directly be added to a ggplot2 object
}
\description{
Imports:
ggplot2
}
\details{
the function uses geom_polygon for the dendrogram, so ... takes all arguements that geom_polygon would also take, such as color, size, alpha, etc.
}
\examples{
library(ggeuler)
library(ggplot2)

set1 <- rownames(subset(mtcars, mpg>18))
set2 <- rownames(subset(mtcars, qsec>18))
set3 <- rownames(subset(mtcars, cyl<5))
carset <- list(highMpg=set1, highQsec=set2, lowCyl=set3)

ggplot2::ggplot() + geom_euler(carset)

}
