% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvenn.R
\name{get_venn_table}
\alias{get_venn_table}
\title{Extract Venn diagram data table from ggvenn plot}
\usage{
get_venn_table(g)
}
\arguments{
\item{g}{A ggplot object created by ggvenn()}
}
\value{
A data frame containing the Venn diagram intersection data
}
\description{
Extract Venn diagram data table from ggvenn plot
}
\examples{
library(ggvenn)
g <- ggvenn(list(A = 1:5, B = 4:9, C = c(2:3, 8:12), D = c(1, 5, 9)))
get_venn_table(g)
}
