% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvenn-package.R
\docType{package}
\name{ggvenn-package}
\alias{ggvenn-package}
\title{ggvenn: Draw Venn Diagram by 'ggplot2'}
\description{
An easy-to-use way to draw pretty Venn diagrams using 'ggplot2'. This package provides functions to create Venn diagrams with customizable colors, labels, and styling options.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yanlinlin82.github.io/ggvenn/}
  \item \url{https://github.com/yanlinlin82/ggvenn}
  \item Report bugs at \url{https://github.com/yanlinlin82/ggvenn/issues}
}

}
\author{
\strong{Maintainer}: Linlin Yan \email{yanlinlin82@gmail.com} (\href{https://orcid.org/0000-0002-4990-6239}{ORCID})

}
\keyword{internal}
