% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvenn.R
\docType{data}
\name{ggvenn}
\alias{ggvenn}
\alias{geom_venn,}
\alias{GeomVenn}
\alias{geom_venn}
\title{Venn Diagram by ggplot}
\format{An object of class \code{GeomVenn} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.}
\usage{
ggvenn(data, columns = NULL)

geom_venn(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ...)

GeomVenn
}
\arguments{
\item{data}{A data.frame or a list as input data.}

\item{columns}{A character vector use as index to select columns/elements.}

\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
The ggplot object to print or save to file
}
\description{
Venn Diagram by ggplot
}
\examples{
library(ggvenn)

# use list as input
a <- list(`Set 1` = c(1, 3, 5, 7, 9),
          `Set 2` = c(1, 5, 9, 13),
          `Set 3` = c(1, 2, 8, 9),
          `Set 4` = c(6, 7, 10, 12))
ggvenn(a, c("Set 1", "Set 2"))
ggvenn(a, c("Set 1", "Set 2", "Set 3"))
ggvenn(a)

# use data.frame as input
d <- tibble(value   = c(1, 2, 3, 5, 6, 7, 8, 9, 10, 12, 13),
            `Set 1` = c(T, F, T, T, F, T, F, T, F,  F,  F),
            `Set 2` = c(T, F, F, T, F, F, F, T, F,  F,  T),
            `Set 3` = c(T, T, F, F, F, F, T, T, F,  F,  F),
            `Set 4` = c(F, F, F, F, T, T, F, F, T,  T,  F))
ggvenn(d, c("Set 1", "Set 2"))
ggvenn(d, c("Set 1", "Set 2", "Set 3"))
ggvenn(d)

# ggplot gramma
ggplot(d) + geom_venn(aes(A = `Set 1`, B = `Set 2`)) + theme_void()
ggplot(d) + geom_venn(aes(A = `Set 1`, B = `Set 2`, C = `Set 3`)) + theme_void()
ggplot(d) + geom_venn(aes(A = `Set 1`, B = `Set 2`, C = `Set 3`, D = `Set 4`)) + theme_void()
}
\keyword{datasets}
