% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_data_frame.R
\name{list_to_data_frame}
\alias{list_to_data_frame}
\title{Utility function for data type conversion.}
\usage{
list_to_data_frame(x)
}
\arguments{
\item{x}{A list of sets.}
}
\value{
A data.frame with logical columns representing sets.
}
\description{
Utility function for data type conversion.
}
\examples{
a <- list(A = 1:5, B = 4:6)
print(a)
list_to_data_frame(a)
}
