% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn.R
\name{plot_venn}
\alias{plot_venn}
\title{Plot Venn or Euler diagram with two sets}
\usage{
plot_venn(
  dfA,
  dfB,
  by = NULL,
  na_matches = "never",
  euler = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dfA}{Data frame}

\item{dfB}{Data frame}

\item{by}{A character vector of variables to join by. If \code{NULL}, a natural join is performed.}

\item{na_matches}{Should \code{NA} and \code{NAN} values count as a match between \code{dfA} and \code{dfB}? Either \code{"na"} or \code{"never"}. If \code{"na"}, they do count as matches and if \code{"never"}, they don’t. (default=\code{"never"})}

\item{euler}{If set to \code{TRUE} a Euler diagram is drawn instead of a Venn diagram (default=\code{FALSE})}

\item{verbose}{If set to \code{TRUE}, logging information is printed out (default=\code{FALSE})}
}
\value{
A ggplot object with the Venn or Euler diagram.
}
\description{
Plot Venn or Euler diagram with two sets
}
\examples{
# Generate sample data
dfA <- data.frame(movie=c("Finding Nemo", "Shrek", "Toy Story"), rating=c(5, 3, 7))
dfB <- data.frame(movie=c("Finding Nemo", "Shrek", "Toy Story", "Inception"), rating=c(5, 3, 7, 8))

# Plot Venn diagram
plot_venn(dfA, dfB)

# Plot Euler diagram
plot_venn(dfA, dfB, euler=TRUE)
}
